Move shared schedule page logic into SharedScheduleBanner
