+++ apps/antalmanac/src/components/SharedScheduleBanner.tsx
@@ -1,0 +1,334 @@
+import { Add, Close } from '@mui/icons-material';
+import { useMediaQuery, useTheme, Stack, Alert, Button, Box, Typography, IconButton } from '@mui/material';
+import type {
+    ScheduleSaveState,
+    ShortCourseSchedule,
+    ShortCourse,
+    RepeatingCustomEvent,
+} from '@packages/antalmanac-types';
+import { useCallback, useEffect, useRef, useState } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
+
+import { importSharedScheduleById, openSnackbar } from '$actions/AppStoreActions';
+import trpc from '$lib/api/trpc';
+import { removeLocalStorageUnsavedActions } from '$lib/localStorage';
+import { getDefaultTerm } from '$lib/termData';
+import AppStore from '$stores/AppStore';
+import { scheduleComponentsToggleStore } from '$stores/ScheduleComponentsToggleStore';
+import { useSessionStore } from '$stores/SessionStore';
+
+interface Props {
+    error: string | null;
+    setError: (error: string | null) => void;
+}
+
+const SharedScheduleBanner = ({ error, setError }: Props) => {
+    const theme = useTheme();
+    const navigate = useNavigate();
+    const { scheduleId } = useParams<{ scheduleId: string }>();
+    const { sessionIsValid } = useSessionStore();
+    const { setOpenLoadingSchedule } = scheduleComponentsToggleStore();
+    const [scheduleName, setScheduleName] = useState<string | null>(null);
+    const hasAttemptedLoadRef = useRef(false);
+    const currentScheduleIdRef = useRef<string | undefined>(scheduleId);
+
+    const isMobileScreen = useMediaQuery(theme.breakpoints.down('sm'));
+
+    useEffect(() => {
+        if (currentScheduleIdRef.current !== scheduleId) {
+            hasAttemptedLoadRef.current = false;
+            currentScheduleIdRef.current = scheduleId;
+        }
+
+        if (hasAttemptedLoadRef.current) {
+            return;
+        }
+
+        const loadSharedSchedule = async () => {
+            if (!scheduleId) {
+                setError('Invalid schedule ID');
+                setOpenLoadingSchedule(false);
+                hasAttemptedLoadRef.current = true;
+                return;
+            }
+
+            hasAttemptedLoadRef.current = true;
+
+            try {
+                setOpenLoadingSchedule(true);
+
+                removeLocalStorageUnsavedActions();
+
+                if (AppStore.getSkeletonMode()) {
+                    AppStore.exitSkeletonMode();
+                }
+
+                const sharedSchedule = await trpc.userData.getSharedSchedule.query({
+                    scheduleId,
+                });
+
+                const scheduleSaveState = {
+                    schedules: [
+                        {
+                            id: sharedSchedule.id,
+                            scheduleName: sharedSchedule.scheduleName,
+                            scheduleNote: sharedSchedule.scheduleNote || '',
+                            courses: sharedSchedule.courses,
+                            customEvents: sharedSchedule.customEvents,
+                        },
+                    ],
+                    scheduleIndex: 0,
+                };
+
+                const loadSuccess = await AppStore.loadSchedule(scheduleSaveState);
+
+                if (!loadSuccess) {
+                    throw new Error('Failed to load shared schedule');
+                }
+                if (AppStore.getCurrentScheduleIndex() !== 0) {
+                    AppStore.changeCurrentSchedule(0);
+                }
+                AppStore.skeletonMode = true;
+                AppStore.emit('skeletonModeChange');
+                setScheduleName(sharedSchedule.scheduleName);
+                setOpenLoadingSchedule(false);
+                setError(null);
+            } catch (err) {
+                console.error('Error loading shared schedule:', err);
+                setError('Failed to load shared schedule. It may not exist or may have been deleted.');
+                setOpenLoadingSchedule(false);
+            }
+        };
+
+        loadSharedSchedule();
+
+        return () => {
+            setOpenLoadingSchedule(false);
+        };
+    }, [scheduleId, setOpenLoadingSchedule, setError]);
+
+    const handleExitSharedSchedule = useCallback(async () => {
+        try {
+            setOpenLoadingSchedule(true);
+
+            if (AppStore.getSkeletonMode()) {
+                AppStore.exitSkeletonMode();
+            }
+
+            if (sessionIsValid) {
+                const sessionToken = useSessionStore.getState().session;
+
+                if (sessionToken) {
+                    const userDataResponse = await trpc.userData.getUserDataWithSession.query({
+                        refreshToken: sessionToken,
+                    });
+                    const scheduleSaveState =
+                        (userDataResponse as { userData?: ScheduleSaveState }).userData ?? userDataResponse;
+
+                    if (scheduleSaveState) {
+                        await AppStore.loadSchedule(scheduleSaveState as ScheduleSaveState);
+                    }
+                }
+            } else {
+                const defaultTerm = getDefaultTerm();
+                const emptyScheduleData = {
+                    scheduleName: defaultTerm.shortName.replaceAll(' ', '-'),
+                    courses: [] as ShortCourse[],
+                    customEvents: [] as RepeatingCustomEvent[],
+                    scheduleNote: '',
+                } as ShortCourseSchedule;
+                const emptySchedule: ScheduleSaveState = {
+                    schedules: [emptyScheduleData],
+                    scheduleIndex: 0,
+                };
+                await AppStore.loadSchedule(emptySchedule);
+            }
+
+            setOpenLoadingSchedule(false);
+            navigate('/');
+        } catch (err) {
+            console.error('Error exiting shared schedule:', err);
+            setOpenLoadingSchedule(false);
+            navigate('/');
+        }
+    }, [navigate, sessionIsValid, setOpenLoadingSchedule]);
+
+    const handleLoadSchedule = useCallback(async (sessionToken?: string) => {
+        if (sessionToken) {
+            const userDataResponse = await trpc.userData.getUserDataWithSession.query({
+                refreshToken: sessionToken,
+            });
+            const scheduleSaveState =
+                (userDataResponse as { userData?: ScheduleSaveState }).userData ?? userDataResponse;
+
+            if (scheduleSaveState) {
+                await AppStore.loadSchedule(scheduleSaveState as ScheduleSaveState);
+            }
+        } else {
+            const defaultTerm = getDefaultTerm();
+            const emptyScheduleData = {
+                scheduleName: defaultTerm.shortName.replaceAll(' ', '-'),
+                courses: [] as ShortCourse[],
+                customEvents: [] as RepeatingCustomEvent[],
+                scheduleNote: '',
+            } as ShortCourseSchedule;
+            const emptySchedule: ScheduleSaveState = {
+                schedules: [emptyScheduleData],
+                scheduleIndex: 0,
+            };
+            await AppStore.loadSchedule(emptySchedule);
+        }
+    }, []);
+
+    const handleAddToMySchedules = useCallback(async () => {
+        if (!scheduleId) {
+            return;
+        }
+
+        try {
+            setOpenLoadingSchedule(true);
+
+            if (AppStore.getSkeletonMode()) {
+                AppStore.exitSkeletonMode();
+            }
+
+            if (sessionIsValid) {
+                const sessionToken = useSessionStore.getState().session;
+                if (!sessionToken) {
+                    throw new Error('No session token available');
+                }
+
+                await handleLoadSchedule(sessionToken);
+                await importSharedScheduleById(scheduleId);
+            } else {
+                const currentSchedules = AppStore.schedule.getScheduleAsSaveState();
+                const sharedSchedule = await trpc.userData.getSharedSchedule.query({ scheduleId });
+                const currentSchedule = currentSchedules.schedules[currentSchedules.scheduleIndex];
+
+                if (currentSchedule && currentSchedule.scheduleName === sharedSchedule.scheduleName) {
+                    const prefixedName = `(shared)-${sharedSchedule.scheduleName}`;
+                    AppStore.renameSchedule(currentSchedules.scheduleIndex, prefixedName);
+                    openSnackbar(
+                        'success',
+                        `Shared schedule "${sharedSchedule.scheduleName}" added to your schedules!`
+                    );
+                } else {
+                    await importSharedScheduleById(scheduleId);
+                }
+            }
+
+            setOpenLoadingSchedule(false);
+            navigate('/');
+        } catch (err) {
+            console.error('Error adding schedule to account:', err);
+            setOpenLoadingSchedule(false);
+            if (AppStore.getSkeletonMode()) {
+                AppStore.exitSkeletonMode();
+            }
+            navigate('/');
+        }
+    }, [scheduleId, sessionIsValid, navigate, setOpenLoadingSchedule, handleLoadSchedule]);
+
+    const handleGoHome = useCallback(async () => {
+        try {
+            setOpenLoadingSchedule(true);
+
+            if (AppStore.getSkeletonMode()) {
+                AppStore.exitSkeletonMode();
+            }
+
+            const sessionToken = useSessionStore.getState().session;
+
+            if (!sessionToken) {
+                throw new Error('No session token available');
+            }
+
+            await handleLoadSchedule(sessionToken);
+
+            setOpenLoadingSchedule(false);
+            navigate('/');
+        } catch (err) {
+            console.error('Error loading user data:', err);
+            setOpenLoadingSchedule(false);
+            navigate('/');
+        }
+    }, [navigate, setOpenLoadingSchedule, handleLoadSchedule]);
+
+    if (error) {
+        return (
+            <>
+                <Alert severity="error">{error}</Alert>
+                <Button variant="contained" onClick={handleGoHome}>
+                    Go Home
+                </Button>
+            </>
+        );
+    }
+
+    return (
+        <Box
+            sx={{
+                px: 2,
+                py: isMobileScreen ? 0.5 : 1,
+                backgroundColor: 'background.paper',
+                borderBottom: 1,
+                borderColor: 'divider',
+            }}
+        >
+            <Stack
+                direction={'row'}
+                justifyContent="space-between"
+                alignItems={'center'}
+                spacing={isMobileScreen ? 1 : 0}
+            >
+                <Typography
+                    variant="h6"
+                    component="h1"
+                    sx={{
+                        fontSize: isMobileScreen ? '1rem' : undefined,
+                        overflow: 'hidden',
+                        textOverflow: 'ellipsis',
+                        display: 'flex',
+                        alignItems: 'center',
+                    }}
+                >
+                    Viewing Shared Schedule:{' '}
+                    <strong style={{ whiteSpace: 'nowrap', marginLeft: '0.25rem' }}>{scheduleName}</strong>
+                </Typography>
+                <Stack direction="row" spacing={1} alignItems="center" sx={{ flexShrink: 0 }}>
+                    {isMobileScreen ? (
+                        <IconButton
+                            aria-label="Add to My Schedules"
+                            onClick={handleAddToMySchedules}
+                            size="small"
+                            sx={{
+                                backgroundColor: 'primary.dark',
+                                color: 'white',
+                                borderRadius: 2,
+                                '&:hover': {
+                                    backgroundColor: 'primary.main',
+                                },
+                            }}
+                        >
+                            <Add />
+                        </IconButton>
+                    ) : (
+                        <Button variant="contained" onClick={handleAddToMySchedules} size="large">
+                            Add to My Schedules
+                        </Button>
+                    )}
+                    <IconButton
+                        aria-label="Exit shared schedule"
+                        onClick={handleExitSharedSchedule}
+                        color="inherit"
+                        size={isMobileScreen ? 'medium' : 'large'}
+                    >
+                        <Close />
+                    </IconButton>
+                </Stack>
+            </Stack>
+        </Box>
+    );
+};
+
+export default SharedScheduleBanner;
